hax: ResolutionUpdate(): Do not resize the crop rectangle
Somewhere between M and N dynamic resolution change in the Exynos decoders broke, causing videos that autoscale - like YouTube on Auto quality - to zoom in, showing only a small portion of the video after it resizes from a low to a high resolution.
While not the best solution, this commit makes it so this bug doesn't happen - as tested in the YouTube app.

Change-Id: I7c686adc8b9219524216df453312151666f88e7d